moved any functions related to input to searchbar (since that's where the button code is as well)
