+++ phillygpt/src/pages/home.jsx
@@ -2,19 +2,8 @@ import React from 'react';
 import SearchBar from '../components/searchbar.jsx';
 import DisplayArea from '../components/displayArea.jsx';
 import Title from '../components/title.jsx';
-import { useState } from 'react';
 
 const HomePage = () => {
-
-  const [userInput, setUserInput] = useState('');
-  
-  const handleInputChange = (event) => {
-    setUserInput(event.target.value);
-  };
-
-  const handleButtonClick = () => {
-    console.log(userInput);
-  };
 
   return (
     <div>
@@ -22,11 +11,7 @@ const HomePage = () => {
       <Title route={'/'} />
 
       {/*SearchBar + Button*/}
-      <SearchBar
-        userInput={userInput}
-        handleInputChange={handleInputChange}
-        handleButtonClick={handleButtonClick}
-      />
+      <SearchBar/>
 
       {/*Display area + Response Box*/}
       <div className="display-area-container">
